+++ ShopGo/OTPViewController+Storyboard/OTPViewController.swift
@@ -9,6 +9,7 @@ import UIKit
 
 class OTPViewController: UIViewController {
     
+    let addDelay: Double = 0.0
     @IBOutlet weak var firstField: UITextField!
     @IBOutlet weak var secondField: UITextField!
     @IBOutlet weak var thirdField: UITextField!
@@ -31,52 +32,5 @@ class OTPViewController: UIViewController {
         createAccountButton.layer.shadowOffset = CGSize(width: 0, height: 5)
         createAccountButton.layer.shadowRadius = 3.0
         createAccountButton.layer.shadowOpacity = 0.3
-    }
-}
-
-extension OTPViewController: UITextFieldDelegate {
-    
-    func textField(_ textField: UITextField, shouldChangeCharactersIn range: NSRange, replacementString string: String) -> Bool {
-        if (textField.text?.count)! < 1 && string.count > 0 {
-            switch textField {
-            case firstField:
-                DispatchQueue.main.asyncAfter(deadline: .now() + 0.1) {
-                    self.secondField.becomeFirstResponder()
-                }
-            case secondField:
-                DispatchQueue.main.asyncAfter(deadline: .now() + 0.1) {
-                    self.thirdField.becomeFirstResponder()
-                }
-            case thirdField:
-                DispatchQueue.main.asyncAfter(deadline: .now() + 0.1) {
-                    self.fourthField.becomeFirstResponder()
-                }
-            case fourthField:
-                if fourthField.text != nil  {
-                    DispatchQueue.main.asyncAfter(deadline: .now() + 0.3) {
-                        self.fourthField.resignFirstResponder()
-                    }
-                }
-            default: break
-            }
-            
-        } else if (textField.text?.count ?? 0) >= 1 {
-            switch textField {
-            case fourthField:
-                DispatchQueue.main.asyncAfter(deadline: .now() + 0.1) {
-                    self.thirdField.becomeFirstResponder()
-                }
-            case thirdField:
-                DispatchQueue.main.asyncAfter(deadline: .now() + 0.1) {
-                    self.secondField.becomeFirstResponder()
-                }
-            case secondField:
-                DispatchQueue.main.asyncAfter(deadline: .now() + 0.1) {
-                    self.firstField.becomeFirstResponder()
-                }
-            default: break
-            }
-        }
-        return true
     }
 }
