+++ ShopGo/ViewController.swift
@@ -14,6 +14,5 @@ class ViewController: UIViewController {
         // Do any additional setup after loading the view.
     }
 
-
 }
